+++ frontend/arrangor-flate/app/routes/$orgnr._manuell-innsending.opprett-krav.innsendingsinformasjon.tsx
@@ -314,7 +314,7 @@ export default function OpprettKravInnsendingsinformasjon() {
                 >
                   {relevanteTilsagn.map((tilsagn) => (
                     <Radio key={tilsagn.id} size="small" value={tilsagn.id}>
-                      <TilsagnDetaljer key={tilsagn.id} tilsagn={tilsagn} />
+                      <TilsagnDetaljer key={tilsagn.id} tilsagn={tilsagn} headingLevel="4" />
                     </Radio>
                   ))}
                 </RadioGroup>
