+++ frontend/arrangor-flate/app/routes/$orgnr.utbetaling.$id.kvittering.tsx
@@ -57,56 +57,53 @@ export default function UtbetalingKvittering() {
   const orgnr = useOrgnrFromUrl();
 
   return (
-    <>
-      <VStack gap="5" className="max-w-[50%] my-5 mx-auto">
-        <PageHeading
-          title={tekster.bokmal.utbetaling.kvittering.headingTitle}
-          tilbakeLenke={{
-            navn: tekster.bokmal.tilbakeTilOversikt,
-            url: pathByOrgnr(orgnr).utbetalinger,
-          }}
-        />
-        <Alert variant="success">{tekster.bokmal.utbetaling.kvittering.successMelding}</Alert>
-        <ExpansionCard defaultOpen aria-label="Kvittering">
-          <ExpansionCard.Header>
-            <ExpansionCard.Title>
-              {tekster.bokmal.utbetaling.kvittering.kvitteringTitle}
-            </ExpansionCard.Title>
-          </ExpansionCard.Header>
-          <ExpansionCard.Content>
-            <VStack gap="2">
-              <BodyShort>
-                {tekster.bokmal.utbetaling.kvittering.mottattAv(mottattTidspunkt)}
-              </BodyShort>
-              <BodyShort>{tekster.bokmal.utbetaling.kvittering.orgnr(orgnr)}</BodyShort>
-              <br />
-              {id && (
-                <BodyLong>
-                  {tekster.bokmal.utbetaling.kvittering.statusLenkeIntro}{" "}
-                  <Link as={ReactRouterLink} to={pathByOrgnr(orgnr).detaljer(id)}>
-                    {tekster.bokmal.utbetaling.kvittering.statusLenkeTekst.toLowerCase()}
-                  </Link>
-                </BodyLong>
-              )}
-            </VStack>
-          </ExpansionCard.Content>
-        </ExpansionCard>
-        <ExpansionCard open aria-label="Kvittering">
-          <ExpansionCard.Header>
-            <ExpansionCard.Title>
-              {tekster.bokmal.utbetaling.kvittering.kontoTitle}
-            </ExpansionCard.Title>
-          </ExpansionCard.Header>
-          <ExpansionCard.Content>
-            <VStack gap="2">
-              <BodyShort weight="semibold">
-                {tekster.bokmal.utbetaling.kvittering.kontonummerRegistrert}
-              </BodyShort>
-              <BodyShort>{kontonummer}</BodyShort>
-            </VStack>
-          </ExpansionCard.Content>
-        </ExpansionCard>
-      </VStack>
-    </>
+    <VStack gap="5" className="max-w-[50%] my-5 mx-auto">
+      <PageHeading
+        title={tekster.bokmal.utbetaling.kvittering.headingTitle}
+        tilbakeLenke={{
+          navn: tekster.bokmal.tilbakeTilOversikt,
+          url: pathByOrgnr(orgnr).utbetalinger,
+        }}
+      />
+      <Alert variant="success">{tekster.bokmal.utbetaling.kvittering.successMelding}</Alert>
+      <ExpansionCard defaultOpen aria-label={tekster.bokmal.utbetaling.kvittering.kvitteringTitle}>
+        <ExpansionCard.Header>
+          <ExpansionCard.Title>
+            {tekster.bokmal.utbetaling.kvittering.kvitteringTitle}
+          </ExpansionCard.Title>
+        </ExpansionCard.Header>
+        <ExpansionCard.Content>
+          <VStack gap="2">
+            <BodyShort>
+              {tekster.bokmal.utbetaling.kvittering.mottattAv(mottattTidspunkt)}
+            </BodyShort>
+            <BodyShort spacing>{tekster.bokmal.utbetaling.kvittering.orgnr(orgnr)}</BodyShort>
+            {id && (
+              <BodyLong>
+                {tekster.bokmal.utbetaling.kvittering.statusLenkeIntro}{" "}
+                <Link as={ReactRouterLink} to={pathByOrgnr(orgnr).detaljer(id)}>
+                  {tekster.bokmal.utbetaling.kvittering.statusLenkeTekst.toLowerCase()}
+                </Link>
+              </BodyLong>
+            )}
+          </VStack>
+        </ExpansionCard.Content>
+      </ExpansionCard>
+      <ExpansionCard open aria-label={tekster.bokmal.utbetaling.kvittering.kontoTitle}>
+        <ExpansionCard.Header>
+          <ExpansionCard.Title>
+            {tekster.bokmal.utbetaling.kvittering.kontoTitle}
+          </ExpansionCard.Title>
+        </ExpansionCard.Header>
+        <ExpansionCard.Content>
+          <VStack gap="2">
+            <BodyShort weight="semibold">
+              {tekster.bokmal.utbetaling.kvittering.kontonummerRegistrert}
+            </BodyShort>
+            <BodyShort>{kontonummer}</BodyShort>
+          </VStack>
+        </ExpansionCard.Content>
+      </ExpansionCard>
+    </VStack>
   );
 }
